+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `consLength`.
 * Add `putManyConsLength`, useful for chunked HTTP encoding.
 * Add `runOnto`
+* Add `Data.Bytes.Chunks.length`
 
 ## 0.3.1.0 -- 2019-11-20